fix: Remove nginx
Brings setup in line with this PR: https://github.com/canonical/maas/commit/ab68c2d7d0847692510a7d29b06e0deeb26dc37a

Since that PR, when I try to run maas-dev-setup, it would fail in-container when trying to stop the nginx service. This is because `make install-dependencies` wouldn't install it as it was removed from the base package list.

This commit fixes this and allows setup to continue.